Add files for professor, student and general api in frontend and separate professor and student thesis proposal page
