+++ code/client/src/apis/generalAPI.js
@@ -1,12 +1,13 @@
 const SERVER_URL = 'http://localhost:3001';
 
 const getUserInfo = async() => {
-    const student = {
+    //TODO: fetch real data
+    const user = { 
         "id": 200001,
         "email": "mario.rossi@studenti.polito.it",	
         "role": "student"
     }
-    return student;
+    return user;
 }
 const API= {getUserInfo};
 export default API;
